fix(formvalidator): remove unnecessary fields param
when creating a custom validator, there is an unncessary param that is required which is the fields
that the custom validation will touch during its execution. This used to be used since it gets
recycled back to the validator to serve as a snapshot validation. Meaning that running the validate
method will validated the value at the time the validator has been created, any further changes
happened after will not be included in the validation. This feature has been removed recently
therefore there is no need for this parameter to exist. Just pass the fields where the error will
appear in and the custom validation method